enforce schema column ordering in data display
